requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
configparser==5.3.0
fastapi==0.95.1
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
lxml==4.9.2
MechanicalSoup==1.2.0
multidict==6.0.4
newsapi-python==0.2.7
openai==0.27.5
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.2
yarl==1.9.2
